feat(NO-US): Add rollup and main bundler
Rollup was created to build flat distributables as efficiently as
possible. It supports ES import/export out of the box which allows us to
take advantage of language (e.g.: tree-shaking).
It's a lot smaller than Webpack, faster and easier to configure. It also
easily supports external plugins to extend its capabilities.