feat(NO-US): Remove polyfills from bundled files.
Is stated in the documentation that the Statful client supports every
browser with ECMAScript 5 support, which means that we don't need to
include the js-polyfill/es5.js.
Most browsers by now support the user-timing api, which means that by
including the usertiming polyfill we are actually hurting the majority
of our users. We can point to the polyfill for those in need to support
older browsers versions.

BREAKING CHANGE: By removing the polyfills we might break support for
developers already relying on them.