chore(NO-US): Use standard .eslintrc.js file
eslint.json is not the standard file to store eslint rules. Although
Grunt allows to specify a file, this is not advantageous for other tools
like editors/IDEs, since those could not automatically find the file.
Also, most of the rules can be replaced with the recommended settings
from eslint.