chore(NO-US): Remove dependencies from bower
Bower is slowly loosing traction and themselves are now recommending
yarn. In order to avoid future problems we can move the devDependencies
to npm and import them directly from node_modules. We can still keep the
bower.json for legacy reasons and in order to avoid problems with older
installations.